requirements.txt
py==1.9.0
pytest==5.4.3
GitPython==3.1.7